made changes to the navbar, landing page, login page, and footer components, created login section component and user icon svg component
