+++ README.md
@@ -19,13 +19,30 @@ Part of description that should be added
 
 ### Product
 - Add Product to Stock
+- Categories
 - Edit Product on Stock
 - Delete Product from Stock
 
 ### Stock
 - Display all Products available
+- Create list view
 
 ### Statistics
 - Based on Popularity
 - Based on Price
 - Displayed graphic
+
+## Bidding system
+- How should this be implemented
+
+##Order cancellations
+- Policy on returns
+  - How long/what circumstances
+- Return method(s)
+
+add terms and agreements radio input to sign up page, required input, we need proof in the database of user having ticked the terms and agreements
+
+landing page - call to action
+sell stuff? go here : buy stuff here
+
+make page to show other user profiles: show name, items selling
