+++ client/package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^13.2.1",
     "axios": "^0.24.0",
     "bootstrap": "^5.1.3",
+    "gsap": "^3.9.1",
     "react": "^17.0.2",
     "react-bootstrap": "^2.0.4",
     "react-dom": "^17.0.2",
