+++ client/src/components/block/Footer/Footer.js
@@ -4,7 +4,7 @@ const Footer = () => {
     return (
         <footer id={"footer"}>
             <div id={"footerAbove"}>
-                <Link to={"/"}>B-Bay</Link>
+                <Link to={"/"}>G-Bay</Link>
                 <Link to={"/about-us"}>About us</Link>
                 <Link to={"/contact"}>Contact</Link>
             </div>
