+++ client/src/components/block/SVGs/UserIconSVG/_UserIconSvg.scss
@@ -1,0 +1,3 @@
+#userIconSvg {
+    shape-rendering: auto;
+}
