+++ client/src/components/block/_index.scss
@@ -1,3 +1,4 @@
 @forward './ProductCard/ProductCard';
 @forward "./Navbar/Navbar";
 @forward "./Footer/Footer";
+@forward "SVGs/UserIconSVG/UserIconSvg";
