+++ client/src/components/container/LoginSection/_LoginSection.scss
@@ -1,0 +1,85 @@
+#loginPage {
+    display: flex;
+    width: 100%;
+    height: 100vh;
+
+    justify-content: center;
+    align-items: center;
+
+    #loginContainer {
+        display: flex;
+        width: 30%;
+        height: 95%;
+
+        border-radius: 36px;
+        border: 1px solid #474044;
+
+        flex-direction: column;
+        place-items: center;
+        text-align: center;
+        justify-content: space-evenly;
+
+        overflow-x: hidden;
+
+        #loginLogo {
+            position: relative;
+            font-size: 6vw;
+            z-index: 1;
+        }
+
+        #loginLogoCropped {
+            position: relative;
+            width: 50vw;
+
+            font-size: 14.5vw;
+            color: #FFFCFF;
+            -webkit-text-stroke: 1px #474044;
+
+            overflow-x: hidden;
+            z-index: 0;
+        }
+
+        #loginForm {
+            position: absolute;
+            display: flex;
+            height: 42.5%;
+            width: 18%;
+
+            flex-direction: column;
+            justify-content: space-evenly;
+
+            z-index: 1;
+
+            input {
+                height: 5vh;
+                font-size: 1rem;
+            }
+
+            #loginFormButtonContainer {
+                display: flex;
+                flex-direction: row;
+                justify-content: center;
+
+                button {
+                    width: 45%;
+                    height: 5vh;
+                    font-size: 1rem;
+                    border-radius: 11px;
+                }
+            }
+        }
+
+        #loginLinks {
+            display: flex;
+            flex-direction: column;
+            justify-content: space-evenly;
+            height: 15%;
+
+            #signUp, #passwordReset {
+                font-size: 1.15rem;
+                text-decoration: none;
+                color: var(--clr-darkgray)
+            }
+        }
+    }
+}
