+++ client/src/components/container/_index.scss
@@ -1,2 +1,3 @@
 @forward './StockSection/StockSection';
 @forward "SingleProductSection/SingleProductSection";
+@forward "LoginSection/LoginSection";
