+++ client/src/style/style.css
@@ -102,6 +102,34 @@ body,
   margin-right: 4vw;
   align-items: center;
   gap: 5vw;
+  height: 100%;
+}
+#navbar #navRight #navDropdown {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+#navbar #navRight #navDropdown #navDropdownButton {
+  border: none;
+  outline: none;
+  background-color: inherit;
+  border-radius: 50%;
+  width: 2.5rem;
+  height: 2.5rem;
+  transform: scale(1.5);
+}
+#navbar #navRight #navDropdown #navDropdownContent {
+  display: none;
+  opacity: 0;
+  background-color: var(--clr-white);
+  border: 1px solid var(--clr-darkgray);
+  border-radius: 11px;
+  min-width: 6vw;
+  min-height: 12vh;
+  transform: translate(0, 90%);
+  position: absolute;
+  z-index: 1;
 }
 #navbar #navRight .navRight {
   font-size: 1.25vw;
@@ -137,6 +165,10 @@ body,
   font-size: 0.75vw;
   justify-content: space-evenly;
   place-items: center;
+}
+
+#userIconSvg {
+  shape-rendering: auto;
 }
 
 #stockSection {
@@ -189,7 +221,7 @@ body,
 
 #loginPage {
   display: flex;
-  width: 100vw;
+  width: 100%;
   height: 100vh;
   justify-content: center;
   align-items: center;
@@ -236,7 +268,7 @@ body,
 #loginPage #loginContainer #loginForm #loginFormButtonContainer {
   display: flex;
   flex-direction: row;
-  justify-content: space-between;
+  justify-content: center;
 }
 #loginPage #loginContainer #loginForm #loginFormButtonContainer button {
   width: 45%;
@@ -244,9 +276,16 @@ body,
   font-size: 1rem;
   border-radius: 11px;
 }
-#loginPage #loginContainer #passwordReset {
+#loginPage #loginContainer #loginLinks {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  height: 15%;
+}
+#loginPage #loginContainer #loginLinks #signUp, #loginPage #loginContainer #loginLinks #passwordReset {
   font-size: 1.15rem;
   text-decoration: none;
+  color: var(--clr-darkgray);
 }
 
 #suPage {
